requirements.txt
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
psutil>=5.9.0
watchdog>=3.0.0
colorama>=0.4.6
pyyaml>=6.0
duckduckgo-search>=4.0
pygetwindow>=0.0.9
rich>=13.0.0
